requirements.txt
python-dotenv
langchain
firecrawl-py==0.0.20
langchain-chroma
langchain-openai
PyQt6